Initial assignment editor skeletons - rename some files for consistency
