+++ demo/d2l-activity-editor/d2l-activity-assignment-editor/d2l-activity-assignment-editor-detail.html
@@ -5,7 +5,7 @@
 		<script src="../../../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 		<script type="module">
 			import '@brightspace-ui/core/components/demo/demo-page.js';
-			import '../../../components/d2l-activity-editor/d2l-activity-assignment-editor/d2l-assignment-editor-detail.js';
+			import '../../../components/d2l-activity-editor/d2l-activity-assignment-editor/d2l-activity-assignment-editor-detail.js';
 		</script>
 		<title>d2l-activity-assignment-editor</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
